requirements.txt
gTTS==2.2.3
opencv_contrib_python==4.5.5.64
oyaml==1.0
Pillow==9.1.1
PySimpleGUI==4.60.0
PyYAML==6.0
rasa==3.1.0
scikit_image==0.19.2
SpeechRecognition==3.8.1
tensorflow==2.7.0
tf_slim==1.1.0
tqdm==4.56.0
